+++ src/components/positionSlider/PositionSlider.tsx
