+++ src/feature/editor/Editor.tsx
@@ -1,5 +1,5 @@
-import Display from './display/Display';
 import GUI from './gui/GUI';
+import Viewport from './viewport/Viewport';
 
 import './Editor.css';
 
@@ -7,7 +7,7 @@ function Editor(): JSX.Element {
   return (
     <div className="Editor">
       <div className="Display">
-        <Display />
+        <Viewport />
       </div>
       <div className="GUI">
         <GUI />
