+++ src/feature/editor/gui/properties/Properties.css
